requirements.txt
streamlit
langchain
langchain-community
google-generativeai
faiss-cpu
sentence-transformers
